split the application header into admin/user side
